Rename folder to comply with Zeitwerk convention
Zeitwerk relies on the convention that each file defines a constant
after the name of the file.

So because the application.rb file here is located in a 'enhancements' folder,
Zeitwerk was expecting a constant Enhancements::Application to be defined there.

Fixes error: 'uninitialized constant Enhancements::Application (NameError)'
