+++ app/models/doorkeeper/application.rb
