+++ config/initializers/load_model_enhancements.rb
@@ -8,8 +8,8 @@
 # Mysql2::Error: Table 'signonotron2_development.supported_permissions' doesn't exist: SHOW FIELDS FROM `supported_permissions`
 #
 unless File.basename($0) == "rake" && ARGV.include?("db:migrate")
-  Dir[Rails.root.join("app/models/enhancements/*.rb")].each do |path|
+  Dir[Rails.root.join("app/models/doorkeeper/*.rb")].each do |path|
     name = File.basename(path, ".rb")
-    require "enhancements/#{name}"
+    require "doorkeeper/#{name}"
   end
 end
