+++ db/migrate/20140114145520_ensure_signin_supported_permission_present.rb
@@ -1,6 +1,6 @@
 class EnsureSigninSupportedPermissionPresent < ActiveRecord::Migration
   def up
-    require "enhancements/application.rb"
+    require "doorkeeper/application.rb"
 
     Doorkeeper::Application.all.each do |application|
       next if application.signin_permission.present?
