+++ db/migrate/20140220150716_most_signin_permissions_should_not_be_delegable.rb
@@ -1,6 +1,6 @@
 class MostSigninPermissionsShouldNotBeDelegable < ActiveRecord::Migration
   def up
-    require "enhancements/application.rb"
+    require "doorkeeper/application.rb"
 
     Doorkeeper::Application.where("name not in ('Content Planner', 'Support')").each do |application|
       application.signin_permission.tap { |p| p.delegatable = false }.save
