+++ db/migrate/20150204132812_add_user_update_supported_permission_to_applications.rb
@@ -1,4 +1,4 @@
-require "enhancements/application"
+require "doorkeeper/application"
 
 class AddUserUpdateSupportedPermissionToApplications < ActiveRecord::Migration
   def change
