+++ db/migrate/20171107153427_add_default_permissions_and_bulk_grant_them_to_all_users.rb
@@ -1,4 +1,4 @@
-require "enhancements/application"
+require "doorkeeper/application"
 
 class AddDefaultPermissionsAndBulkGrantThemToAllUsers < ActiveRecord::Migration
   def up
