+++ db/migrate/20180131120100_set_signin_as_default_permission_for_asset_manager_and_grant_to_all_users.rb
@@ -1,4 +1,4 @@
-require "enhancements/application"
+require "doorkeeper/application"
 
 class SetSigninAsDefaultPermissionForAssetManagerAndGrantToAllUsers < ActiveRecord::Migration[5.1]
   def up
